Morepork-locator: find moreporks by looking for tonal patterns
The canonical morepork call consists of two downward arching
syllables, with the second syllable usually being lower and fainter than
the first. Like this ⏜ ︵ (utf-8), but with more of a drop in tone at the
end of each syllable.

The morepork-locator distinguishes these calls by looking for pure tones
within the morepork frequency band, and comparing their timings to a
template.  It does quite well in noisy environments, but is not very good
at detecting lazy moreporks who hardly pronounce their second syllable.