Script that estimates and stores the intensity of each call
The calls are stored as a series of JSON one-liners. If you imagine
the following all being on one line:

["AI137_BIRM_120203_052707-09+60.wav",
    [0.25, 1.0, 3.894248141223105],
    [4.55, 5.35, 4.376498257003673]]

you'll see it notes two calls in "AI137_BIRM_120203_052707-09+60.wav",
each with a start time, end time, and intensity score, respectively.

To calculate the intensity, the spectrogram power values in the
relevant band of frequencies (set via -L and -H command line
arguments) during the call are sorted and the 5th and 95th centiles
are taken. The score is the log ratio between the two.

In thwe above example, the first call (starting 0.25 seconds in with a
score around 3.9) is barely audible, and the second one is faint. This
suggests a threshold of about 4.0 might work well for moreporks.